+++ sdk/cosmos/azure-mgmt-cosmosdb/azure/mgmt/cosmosdb/_cosmos_db_management_client.py
@@ -32,11 +32,13 @@ from .operations import CollectionPartitionRegionOperations
 from .operations import CollectionPartitionOperations
 from .operations import PartitionKeyRangeIdOperations
 from .operations import PartitionKeyRangeIdRegionOperations
+from .operations import GraphResourcesOperations
 from .operations import SqlResourcesOperations
 from .operations import MongoDBResourcesOperations
 from .operations import TableResourcesOperations
 from .operations import CassandraResourcesOperations
 from .operations import GremlinResourcesOperations
+from .operations import CosmosDBManagementClientOperationsMixin
 from .operations import NotebookWorkspacesOperations
 from .operations import PrivateEndpointConnectionsOperations
 from .operations import PrivateLinkResourcesOperations
@@ -47,10 +49,13 @@ from .operations import RestorableSqlResourcesOperations
 from .operations import RestorableMongodbDatabasesOperations
 from .operations import RestorableMongodbCollectionsOperations
 from .operations import RestorableMongodbResourcesOperations
+from .operations import CassandraClustersOperations
+from .operations import CassandraDataCentersOperations
+from .operations import ServiceOperations
 from . import models
 
 
-class CosmosDBManagementClient(object):
+class CosmosDBManagementClient(CosmosDBManagementClientOperationsMixin):
     """Azure Cosmos DB Database Service Resource Provider REST API.
 
     :ivar database_accounts: DatabaseAccountsOperations operations
@@ -79,6 +84,8 @@ class CosmosDBManagementClient(object):
     :vartype partition_key_range_id: azure.mgmt.cosmosdb.operations.PartitionKeyRangeIdOperations
     :ivar partition_key_range_id_region: PartitionKeyRangeIdRegionOperations operations
     :vartype partition_key_range_id_region: azure.mgmt.cosmosdb.operations.PartitionKeyRangeIdRegionOperations
+    :ivar graph_resources: GraphResourcesOperations operations
+    :vartype graph_resources: azure.mgmt.cosmosdb.operations.GraphResourcesOperations
     :ivar sql_resources: SqlResourcesOperations operations
     :vartype sql_resources: azure.mgmt.cosmosdb.operations.SqlResourcesOperations
     :ivar mongo_db_resources: MongoDBResourcesOperations operations
@@ -109,6 +116,12 @@ class CosmosDBManagementClient(object):
     :vartype restorable_mongodb_collections: azure.mgmt.cosmosdb.operations.RestorableMongodbCollectionsOperations
     :ivar restorable_mongodb_resources: RestorableMongodbResourcesOperations operations
     :vartype restorable_mongodb_resources: azure.mgmt.cosmosdb.operations.RestorableMongodbResourcesOperations
+    :ivar cassandra_clusters: CassandraClustersOperations operations
+    :vartype cassandra_clusters: azure.mgmt.cosmosdb.operations.CassandraClustersOperations
+    :ivar cassandra_data_centers: CassandraDataCentersOperations operations
+    :vartype cassandra_data_centers: azure.mgmt.cosmosdb.operations.CassandraDataCentersOperations
+    :ivar service: ServiceOperations operations
+    :vartype service: azure.mgmt.cosmosdb.operations.ServiceOperations
     :param credential: Credential needed for the client to connect to Azure.
     :type credential: ~azure.core.credentials.TokenCredential
     :param subscription_id: The ID of the target subscription.
@@ -161,6 +174,8 @@ class CosmosDBManagementClient(object):
             self._client, self._config, self._serialize, self._deserialize)
         self.partition_key_range_id_region = PartitionKeyRangeIdRegionOperations(
             self._client, self._config, self._serialize, self._deserialize)
+        self.graph_resources = GraphResourcesOperations(
+            self._client, self._config, self._serialize, self._deserialize)
         self.sql_resources = SqlResourcesOperations(
             self._client, self._config, self._serialize, self._deserialize)
         self.mongo_db_resources = MongoDBResourcesOperations(
@@ -190,6 +205,12 @@ class CosmosDBManagementClient(object):
         self.restorable_mongodb_collections = RestorableMongodbCollectionsOperations(
             self._client, self._config, self._serialize, self._deserialize)
         self.restorable_mongodb_resources = RestorableMongodbResourcesOperations(
+            self._client, self._config, self._serialize, self._deserialize)
+        self.cassandra_clusters = CassandraClustersOperations(
+            self._client, self._config, self._serialize, self._deserialize)
+        self.cassandra_data_centers = CassandraDataCentersOperations(
+            self._client, self._config, self._serialize, self._deserialize)
+        self.service = ServiceOperations(
             self._client, self._config, self._serialize, self._deserialize)
 
     def _send_request(self, http_request, **kwargs):