+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_identity_documents_async.py
@@ -8,18 +8,18 @@ import pytest
 import functools
 from io import BytesIO
 from datetime import date, time
-from azure.core.exceptions import ClientAuthenticationError, ServiceRequestError, HttpResponseError
+from azure.core.exceptions import ServiceRequestError, HttpResponseError
 from azure.core.credentials import AzureKeyCredential
 from azure.ai.formrecognizer._generated.models import AnalyzeOperationResult
 from azure.ai.formrecognizer._response_handlers import prepare_prebuilt_models
-from azure.ai.formrecognizer.aio import FormRecognizerClient
+from azure.ai.formrecognizer.aio import FormRecognizerClient, DocumentAnalysisClient
 from azure.ai.formrecognizer import FormContentType, FormRecognizerApiVersion
 from asynctestcase import AsyncFormRecognizerTest
 from preparers import FormRecognizerPreparer
 from preparers import GlobalClientPreparer as _GlobalClientPreparer
 
-
-GlobalClientPreparer = functools.partial(_GlobalClientPreparer, FormRecognizerClient)
+DocumentAnalysisClientPreparer = functools.partial(_GlobalClientPreparer, DocumentAnalysisClient)
+FormRecognizerClientPreparer = functools.partial(_GlobalClientPreparer, FormRecognizerClient)
 
 
 class TestIdDocumentsAsync(AsyncFormRecognizerTest):
@@ -34,37 +34,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
                 poller = await client.begin_recognize_identity_documents(myfile)
 
     @FormRecognizerPreparer()
-    async def test_authentication_bad_key(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        client = FormRecognizerClient(formrecognizer_test_endpoint, AzureKeyCredential("xxxx"))
-        with self.assertRaises(ClientAuthenticationError):
-            async with client:
-                poller = await client.begin_recognize_identity_documents(b"xx", content_type="image/jpeg")
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    async def test_passing_enum_content_type(self, client):
-        with open(self.identity_document_license_jpg, "rb") as fd:
-            myfile = fd.read()
-        async with client: 
-            poller = await client.begin_recognize_identity_documents(
-                myfile,
-                content_type=FormContentType.IMAGE_JPEG
-            )
-            result = await poller.result()
-        self.assertIsNotNone(result)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    async def test_damaged_file_passed_as_bytes(self, client):
-        damaged_pdf = b"\x25\x50\x44\x46\x55\x55\x55"  # still has correct bytes to be recognized as PDF
-        with self.assertRaises(HttpResponseError):
-            async with client:
-                poller = await client.begin_recognize_identity_documents(
-                    damaged_pdf
-                )
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     async def test_damaged_file_bytes_fails_autodetect_content_type(self, client):
         damaged_pdf = b"\x50\x44\x46\x55\x55\x55"  # doesn't match any magic file numbers
         with self.assertRaises(ValueError):
@@ -74,17 +44,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
                 )
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    async def test_damaged_file_passed_as_bytes_io(self, client):
-        damaged_pdf = BytesIO(b"\x25\x50\x44\x46\x55\x55\x55")  # still has correct bytes to be recognized as PDF
-        with self.assertRaises(HttpResponseError):
-            async with client:
-                poller = await client.begin_recognize_identity_documents(
-                    damaged_pdf
-                )
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     async def test_damaged_file_bytes_io_fails_autodetect(self, client):
         damaged_pdf = BytesIO(b"\x50\x44\x46\x55\x55\x55")  # doesn't match any magic file numbers
         with self.assertRaises(ValueError):
@@ -94,19 +54,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
                 )
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    async def test_blank_page(self, client):
-        with open(self.blank_pdf, "rb") as fd:
-            blank = fd.read()
-        async with client:
-            poller = await client.begin_recognize_identity_documents(
-                blank
-            )
-            result = await poller.result()
-        self.assertIsNotNone(result)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     async def test_passing_bad_content_type_param_passed(self, client):
         with open(self.identity_document_license_jpg, "rb") as fd:
             myfile = fd.read()
@@ -118,14 +66,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
                 )
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    async def test_passing_unsupported_url_content_type(self, client):
-        with self.assertRaises(TypeError):
-            async with client:
-                poller = await client.begin_recognize_identity_documents("https://badurl.jpg", content_type="application/json")
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     async def test_auto_detect_unsupported_stream_content(self, client):
         with open(self.unsupported_content_py, "rb") as fd:
             myfile = fd.read()
@@ -137,7 +78,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
                 )
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     async def test_identity_document_stream_transform_jpg(self, client):
         responses = []
 
@@ -176,55 +117,53 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
         self.assertFormPagesTransformCorrect(id_document.pages, read_results, page_results)
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @DocumentAnalysisClientPreparer()
     async def test_identity_document_jpg_passport(self, client):
         with open(self.identity_document_passport_jpg, "rb") as fd:
             id_document = fd.read()
 
         async with client:
-            poller = await client.begin_recognize_identity_documents(id_document)
+            poller = await client.begin_analyze_document("prebuilt-idDocument", id_document)
 
             result = await poller.result()
-            self.assertEqual(len(result), 1)
+            assert len(result.documents) == 1
         
-            id_document = result[0]
-            # check dict values
+            id_document = result.documents[0]
 
             passport = id_document.fields.get("MachineReadableZone").value
-            self.assertEqual(passport["LastName"].value, "MARTIN")
-            self.assertEqual(passport["FirstName"].value, "SARAH")
-            self.assertEqual(passport["DocumentNumber"].value, "ZE000509")
-            self.assertEqual(passport["DateOfBirth"].value, date(1985,1,1))
-            self.assertEqual(passport["DateOfExpiration"].value, date(2023,1,14))
-            self.assertEqual(passport["Sex"].value, "F")
-            self.assertEqual(passport["CountryRegion"].value, "CAN")
+            assert passport["LastName"].value == "MARTIN"
+            assert passport["FirstName"].value == "SARAH"
+            assert passport["DocumentNumber"].value == "ZE000509"
+            assert passport["DateOfBirth"].value == date(1985,1,1)
+            assert passport["DateOfExpiration"].value == date(2023,1,14)
+            assert passport["Sex"].value == "F"
+            assert passport["CountryRegion"].value == "CAN"
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @DocumentAnalysisClientPreparer()
     async def test_identity_document_jpg(self, client):
         with open(self.identity_document_license_jpg, "rb") as fd:
             id_document = fd.read()
 
         async with client:
-            poller = await client.begin_recognize_identity_documents(id_document)
+            poller = await client.begin_analyze_document("prebuilt-idDocument", id_document)
 
             result = await poller.result()
-        self.assertEqual(len(result), 1)
-        id_document = result[0]
-        # check dict values
+        assert len(result.documents) == 1
+        id_document = result.documents[0]
 
-        self.assertEqual(id_document.fields.get("LastName").value, "TALBOT")
-        self.assertEqual(id_document.fields.get("FirstName").value, "LIAM R.")
-        self.assertEqual(id_document.fields.get("DocumentNumber").value, "WDLABCD456DG")
-        self.assertEqual(id_document.fields.get("DateOfBirth").value, date(1958,1,6))
-        self.assertEqual(id_document.fields.get("DateOfExpiration").value, date(2020,8,12))
-        self.assertEqual(id_document.fields.get("Sex").value, "M")
-        self.assertEqual(id_document.fields.get("Address").value, "123 STREET ADDRESS YOUR CITY WA 99999-1234")
-        self.assertEqual(id_document.fields.get("CountryRegion").value, "USA")
-        self.assertEqual(id_document.fields.get("Region").value, "Washington")
+        assert id_document.fields.get("LastName").value == "TALBOT"
+        assert id_document.fields.get("FirstName").value == "LIAM R."
+        assert id_document.fields.get("DocumentNumber").value == "WDLABCD456DG"
+        assert id_document.fields.get("DateOfBirth").value == date(1958,1,6)
+        assert id_document.fields.get("DateOfExpiration").value == date(2020,8,12)
+        assert id_document.fields.get("Sex").value == "M"
+        assert id_document.fields.get("Address").value == "123 STREET ADDRESS YOUR CITY WA 99999-1234"
+        assert id_document.fields.get("CountryRegion").value == "USA"
+        assert id_document.fields.get("Region").value == "Washington"
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     async def test_identity_document_jpg_include_field_elements(self, client):
         with open(self.identity_document_license_jpg, "rb") as fd:
             id_document = fd.read()
@@ -247,7 +186,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
                 self.assertFieldElementsHasValues(field.value_data.field_elements, id_document.page_range.first_page_number)
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     @pytest.mark.live_test_only
     async def test_identity_document_continuation_token(self, client):
         with open(self.identity_document_license_jpg, "rb") as fd:
@@ -261,7 +200,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
             await initial_poller.wait()  # necessary so azure-devtools doesn't throw assertion error
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer(client_kwargs={"api_version": FormRecognizerApiVersion.V2_0})
+    @FormRecognizerClientPreparer(client_kwargs={"api_version": FormRecognizerApiVersion.V2_0})
     async def test_identity_document_v2(self, client):
         with open(self.identity_document_license_jpg, "rb") as fd:
             id_document = fd.read()
@@ -271,7 +210,7 @@ class TestIdDocumentsAsync(AsyncFormRecognizerTest):
         assert "Method 'begin_recognize_identity_documents' is only available for API version V2_1 and up" in str(e.value)
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     async def test_pages_kwarg_specified(self, client):
         with open(self.identity_document_license_jpg, "rb") as fd:
             id_document = fd.read()