+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_identity_documents_from_url.py
@@ -6,18 +6,17 @@
 
 import pytest
 import functools
-from io import BytesIO
-from datetime import date, time
-from azure.core.exceptions import ClientAuthenticationError, ServiceRequestError, HttpResponseError
+from datetime import date
 from azure.core.credentials import AzureKeyCredential
 from azure.ai.formrecognizer._generated.models import AnalyzeOperationResult
 from azure.ai.formrecognizer._response_handlers import prepare_prebuilt_models
-from azure.ai.formrecognizer import FormRecognizerClient, FormContentType, FormRecognizerApiVersion
+from azure.ai.formrecognizer import FormRecognizerClient, FormRecognizerApiVersion, DocumentAnalysisClient
 from testcase import FormRecognizerTest
 from preparers import GlobalClientPreparer as _GlobalClientPreparer
 from preparers import FormRecognizerPreparer
 
-GlobalClientPreparer = functools.partial(_GlobalClientPreparer, FormRecognizerClient)
+DocumentAnalysisClientPreparer = functools.partial(_GlobalClientPreparer, DocumentAnalysisClient)
+FormRecognizerClientPreparer = functools.partial(_GlobalClientPreparer, FormRecognizerClient)
 
 
 class TestIdDocumentsFromUrl(FormRecognizerTest):
@@ -35,34 +34,7 @@ class TestIdDocumentsFromUrl(FormRecognizerTest):
         self.assertEqual(poller2._polling_method._timeout, 7)  # goes back to client default
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_identity_document_encoded_url(self, client):
-        try:
-            poller = client.begin_recognize_identity_documents_from_url("https://fakeuri.com/blank%20space")
-        except HttpResponseError as e:
-            self.assertIn("https://fakeuri.com/blank%20space", e.response.request.body)
-
-    @FormRecognizerPreparer()
-    def test_authentication_bad_key(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        client = FormRecognizerClient(formrecognizer_test_endpoint, AzureKeyCredential("xxxx"))
-        with self.assertRaises(ClientAuthenticationError):
-            poller = client.begin_recognize_identity_documents_from_url(self.identity_document_url_jpg)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_identity_document_bad_url(self, client):
-        with self.assertRaises(HttpResponseError):
-            poller = client.begin_recognize_identity_documents_from_url("https://badurl.jpg")
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_identity_document_url_pass_stream(self, client):
-        with open(self.identity_document_license_jpg, "rb") as id_document:
-            with self.assertRaises(HttpResponseError):
-                poller = client.begin_recognize_identity_documents_from_url(id_document)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     def test_identity_document_url_transform_jpg(self, client):
         responses = []
 
@@ -97,47 +69,47 @@ class TestIdDocumentsFromUrl(FormRecognizerTest):
         self.assertFormPagesTransformCorrect(id_document.pages, read_results, page_results)
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @DocumentAnalysisClientPreparer()
     def test_identity_document_jpg_passport(self, client):
-        poller = client.begin_recognize_identity_documents_from_url(self.identity_document_url_jpg_passport)
+        poller = client.begin_analyze_document_from_url("prebuilt-idDocument", self.identity_document_url_jpg_passport)
 
         result = poller.result()
-        self.assertEqual(len(result), 1)
+        assert len(result.documents) == 1
     
-        id_document = result[0]
+        id_document = result.documents[0]
         # check dict values
 
         passport = id_document.fields.get("MachineReadableZone").value
-        self.assertEqual(passport["LastName"].value, "MARTIN")
-        self.assertEqual(passport["FirstName"].value, "SARAH")
-        self.assertEqual(passport["DocumentNumber"].value, "ZE000509")
-        self.assertEqual(passport["DateOfBirth"].value, date(1985,1,1))
-        self.assertEqual(passport["DateOfExpiration"].value, date(2023,1,14))
-        self.assertEqual(passport["Sex"].value, "F")
-        self.assertEqual(passport["CountryRegion"].value, "CAN")
+        assert passport["LastName"].value == "MARTIN"
+        assert passport["FirstName"].value == "SARAH"
+        assert passport["DocumentNumber"].value == "ZE000509"
+        assert passport["DateOfBirth"].value == date(1985,1,1)
+        assert passport["DateOfExpiration"].value == date(2023,1,14)
+        assert passport["Sex"].value == "F"
+        assert passport["CountryRegion"].value == "CAN"
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @DocumentAnalysisClientPreparer()
     def test_identity_document_jpg(self, client):
-        poller = client.begin_recognize_identity_documents_from_url(self.identity_document_url_jpg)
+        poller = client.begin_analyze_document_from_url("prebuilt-idDocument", self.identity_document_url_jpg)
 
         result = poller.result()
-        self.assertEqual(len(result), 1)
-        id_document = result[0]
+        assert len(result.documents) == 1
+        id_document = result.documents[0]
 
         # check dict values
-        self.assertEqual(id_document.fields.get("LastName").value, "TALBOT")
-        self.assertEqual(id_document.fields.get("FirstName").value, "LIAM R.")
-        self.assertEqual(id_document.fields.get("DocumentNumber").value, "WDLABCD456DG")
-        self.assertEqual(id_document.fields.get("DateOfBirth").value, date(1958,1,6))
-        self.assertEqual(id_document.fields.get("DateOfExpiration").value, date(2020,8,12))
-        self.assertEqual(id_document.fields.get("Sex").value, "M")
-        self.assertEqual(id_document.fields.get("Address").value, "123 STREET ADDRESS YOUR CITY WA 99999-1234")
-        self.assertEqual(id_document.fields.get("CountryRegion").value, "USA")
-        self.assertEqual(id_document.fields.get("Region").value, "Washington")
+        assert id_document.fields.get("LastName").value == "TALBOT"
+        assert id_document.fields.get("FirstName").value == "LIAM R."
+        assert id_document.fields.get("DocumentNumber").value == "WDLABCD456DG"
+        assert id_document.fields.get("DateOfBirth").value == date(1958,1,6)
+        assert id_document.fields.get("DateOfExpiration").value == date(2020,8,12)
+        assert id_document.fields.get("Sex").value == "M"
+        assert id_document.fields.get("Address").value == "123 STREET ADDRESS YOUR CITY WA 99999-1234"
+        assert id_document.fields.get("CountryRegion").value == "USA"
+        assert id_document.fields.get("Region").value == "Washington"
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     def test_identity_document_jpg_include_field_elements(self, client):
         poller = client.begin_recognize_identity_documents_from_url(self.identity_document_url_jpg, include_field_elements=True)
 
@@ -157,7 +129,7 @@ class TestIdDocumentsFromUrl(FormRecognizerTest):
                 self.assertFieldElementsHasValues(field.value_data.field_elements, id_document.page_range.first_page_number)
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     @pytest.mark.live_test_only
     def test_identity_document_continuation_token(self, client):
         initial_poller = client.begin_recognize_identity_documents_from_url(self.identity_document_url_jpg)
@@ -168,14 +140,14 @@ class TestIdDocumentsFromUrl(FormRecognizerTest):
         initial_poller.wait()  # necessary so azure-devtools doesn't throw assertion error
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer(client_kwargs={"api_version": FormRecognizerApiVersion.V2_0})
+    @FormRecognizerClientPreparer(client_kwargs={"api_version": FormRecognizerApiVersion.V2_0})
     def test_identity_document_v2(self, client):
         with pytest.raises(ValueError) as e:
             client.begin_recognize_identity_documents_from_url(self.identity_document_url_jpg)
         assert "Method 'begin_recognize_identity_documents_from_url' is only available for API version V2_1 and up" in str(e.value)
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
+    @FormRecognizerClientPreparer()
     def test_pages_kwarg_specified(self, client):
         poller = client.begin_recognize_identity_documents_from_url(self.identity_document_url_jpg, pages=["1"])
         assert '1' == poller._polling_method._initial_response.http_response.request.query['pages']